ref: separate cli parts
